requirements.txt
pymongo
oauth2client
gspread
googletrans
textblob
fuzzywuzzy
flask
dnspython
PyYAML
python-dotenv
python-Levenshtein
